pyproject.toml
[project]
name = "car"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fitz>=0.0.1.dev2",
    "fritz>=2.3.5",
    "frontend>=0.0.3",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "notebook>=7.5.3",
    "opencv-python>=4.13.0.92",
    "pdf2image>=1.17.0",
    "pillow>=10.2.0,<11.0.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "structocr>=1.1.4",
    "surya-ocr==0.13.1",
    "transformers>=4.46.0,<4.49.0",
]
